refactor(#110): split body_model.py into body_anthropometrics + thin assembler
Extract BodyModelSpec, _LEG_HEIGHT_FRACTION, and _seg() into
body_anthropometrics.py (anthropometric data and lookup layer).
body_model.py is now a thin MJCF assembly orchestrator that imports
from body_anthropometrics; public API and all imports remain unchanged.
